requirements.txt
fastapi
uvicorn
requests
pypdf2
python-docx